# perf(833): Do not clone query plan during execution