--- a/crates/core/src/db/cursor.rs
+++ b/crates/core/src/db/cursor.rs
@@ -29,12 +29,12 @@ impl<'a> TableCursor<'a> {
 /// A relational iterator wrapping a storage level index iterator.
 /// A relational iterator returns [RelValue]s whereas storage iterators return [DataRef]s.
 pub struct IndexCursor<'a, R: RangeBounds<AlgebraicValue>> {
-    pub table: DbTable,
+    pub table: &'a DbTable,
     pub iter: IterByColRange<'a, R>,
 }
 
 impl<'a, R: RangeBounds<AlgebraicValue>> IndexCursor<'a, R> {
-    pub fn new(table: DbTable, iter: IterByColRange<'a, R>) -> Result<Self, DBError> {
+    pub fn new(table: &'a DbTable, iter: IterByColRange<'a, R>) -> Result<Self, DBError> {
         Ok(Self { table, iter })
     }
 }