--- a/crates/vm/src/eval.rs
+++ b/crates/vm/src/eval.rs
@@ -1,5 +1,3 @@
-use std::sync::Arc;
-
 use crate::errors::ErrorVm;
 use crate::expr::{Code, SourceSet};
 use crate::expr::{Expr, Query};
@@ -7,7 +5,8 @@ use crate::iterators::RelIter;
 use crate::program::ProgramVm;
 use crate::rel_ops::RelOps;
 use crate::relation::RelValue;
-use spacetimedb_sats::relation::{FieldExpr, Relation};
+use spacetimedb_sats::relation::{FieldExprRef, Relation};
+use std::sync::Arc;
 
 pub type IterRows<'a> = dyn RelOps<'a> + 'a;
 
@@ -18,7 +17,7 @@ pub type IterRows<'a> = dyn RelOps<'a> + 'a;
 /// so the `query` cannot refer to the same `SourceId` multiple times.
 pub fn build_query<'a>(
     mut result: Box<IterRows<'a>>,
-    query: Vec<Query>,
+    query: &'a [Query],
     sources: &mut SourceSet,
 ) -> Result<Box<IterRows<'a>>, ErrorVm> {
     for q in query {
@@ -47,10 +46,8 @@ pub fn build_query<'a>(
             }
             Query::JoinInner(q) => {
                 //Pick the smaller set to be at the left
-                let col_lhs = FieldExpr::Name(q.col_lhs);
-                let col_rhs = FieldExpr::Name(q.col_rhs);
-                let key_lhs = col_lhs.clone();
-                let key_rhs = col_rhs.clone();
+                let col_lhs = FieldExprRef::Name(&q.col_lhs);
+                let col_rhs = FieldExprRef::Name(&q.col_rhs);
                 let row_rhs = q.rhs.source.row_count();
 
                 let head = q.rhs.source.head().clone();
@@ -65,7 +62,7 @@ pub fn build_query<'a>(
                     todo!("How pass the db iter?")
                 };
 
-                let rhs = build_query(rhs, q.rhs.query, sources)?;
+                let rhs = build_query(rhs, &q.rhs.query, sources)?;
 
                 let lhs = result;
                 let key_lhs_header = lhs.head().clone();
@@ -76,11 +73,11 @@ pub fn build_query<'a>(
                 let iter = lhs.join_inner(
                     rhs,
                     Arc::new(col_lhs_header.extend(&col_rhs_header)),
-                    move |row| Ok(row.get(&key_lhs, &key_lhs_header)?.into_owned().into()),
-                    move |row| Ok(row.get(&key_rhs, &key_rhs_header)?.into_owned().into()),
+                    move |row| Ok(row.get(col_lhs, &key_lhs_header)?.into_owned().into()),
+                    move |row| Ok(row.get(col_rhs, &key_rhs_header)?.into_owned().into()),
                     move |l, r| {
-                        let l = l.get(&col_lhs, &col_lhs_header)?;
-                        let r = r.get(&col_rhs, &col_rhs_header)?;
+                        let l = l.get(col_lhs, &col_lhs_header)?;
+                        let r = r.get(col_rhs, &col_rhs_header)?;
                         Ok(l == r)
                     },
                     move |l, r| l.extend(r),
@@ -95,7 +92,7 @@ pub fn build_query<'a>(
 /// Execute the code
 pub fn eval<P: ProgramVm>(p: &mut P, code: Code, sources: &mut SourceSet) -> Code {
     match code {
-        Code::Value(_) => code.clone(),
+        c @ (Code::Value(_) | Code::Halt(_) | Code::Table(_)) => c,
         Code::Block(lines) => {
             let mut result = Vec::with_capacity(lines.len());
             for x in lines {
@@ -113,8 +110,6 @@ pub fn eval<P: ProgramVm>(p: &mut P, code: Code, sources: &mut SourceSet) -> Cod
         }
         Code::Crud(q) => p.eval_query(q, sources).unwrap_or_else(|err| Code::Halt(err.into())),
         Code::Pass => Code::Pass,
-        Code::Halt(_) => code,
-        Code::Table(_) => code,
     }
 }