… crates/core/src/vm.rs
… crates/vm/src/rel_ops.rs
… crates/vm/src/relation.rs