# Description of Changes

Fixes #833 

Perf numbers on my machine:

```
Benchmarking incr-join: Collecting 100 samples in estimated 5.01
incr-join               time:   [5.0575 µs 5.0914 µs 5.1521 µs]
                        change: [-19.535% -18.714% -17.870%] (p = 0.00 < 0.05)
                        Performance has improved.

Benchmarking query-indexes-multi: Collecting 100 samples in esti
query-indexes-multi     time:   [1.0330 µs 1.0387 µs 1.0457 µs]
                        change: [-13.574% -12.820% -12.213%] (p = 0.00 < 0.05)
                        Performance has improved.
```

# API and ABI breaking changes

None

# Expected complexity level and risk

1

# Testing

No changes semantically; should be covered by existing tests.