@joshua-spacetime 2024-03-15
Not sure what's going on here
```
full-join               time:   [282.30 µs 283.82 µs 286.06 µs]
                        change: [+11.033% +11.400% +11.830%] (p = 0.00 < 0.05)
                        Performance has regressed.
```